% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_complexes.R
\name{compute_complexes}
\alias{compute_complexes}
\title{Creates network representations of the expression space of a patient cohort across grid of Mapper parameters}
\usage{
compute_complexes(
  TDAmut_object,
  var_threshold = 4500,
  filter_method = "UMAP",
  k = 30,
  min_interval = 10,
  max_interval = 80,
  interval_step = 10,
  min_percent_overlap = 20,
  max_percent_overlap = 90,
  percent_step = 10,
  num_bins = 10,
  recompute = FALSE
)
}
\arguments{
\item{TDAmut_object}{TDAmut class with expression data}

\item{var_threshold}{number of most variable genes to use from input expression table. By default is 4500.}

\item{filter_method}{filter function to create embedding used by TDAmapper. Includes UMAP, KNN, and PCA. By default is set to UMAP.}

\item{k}{if KNN filter function is used, specifies number of nearest neighbors. By default is 30}

\item{min_interval}{minimum number of intervals in codomain of filter function. By default is 10.}

\item{max_interval}{maximum number of intervals in codomain of filter function}

\item{interval_step}{step size between intervals in grid of Mapper parameters. By default is 10.}

\item{min_percent_overlap}{minimum percentage overlap between intervals. By default is 20.}

\item{max_percent_overlap}{maximum percentage overlap between intervals}

\item{percent_step}{step size between percentages in grid of Mapper parameters. By default is 10.}

\item{num_bins}{an integer controlling clustering within the same level set. By default is 10.}

\item{recompute}{if TRUE, recomputes nerve complexes without samples falling short of threshold mutational load identified in compute_mut_load.}
}
\value{
Returns a TDAmut object populated with nerve complexes
}
\description{
Creates network representations of the expression space of a patient cohort across grid of Mapper parameters
}
