% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_gene_localization.R
\name{compute_gene_localization}
\alias{compute_gene_localization}
\title{Computes localization of nonsynonymously mutated genes passing thresholds of filter_genes}
\usage{
compute_gene_localization(
  TDAmut_object,
  negative_correlations_threshold = 0.8,
  num_permutations = 10000,
  num_cores = 5,
  seed = 121
)
}
\arguments{
\item{TDAmut_object}{object of class TDAmut with expression data, mutation data, nerve complexes}

\item{num_permutations}{number of permutations used to create null distribution when assessing localization of mutations in nerve complexes. By default is 10,000}

\item{num_cores}{number of cores to be used in computation. By default is 1.}

\item{seed}{integer specifying the seed used to initialize the generator of permutations. By default is 121}

\item{neg_correlations_threshold}{q value theshold to remove genes displaying correlations between mutation and expression data. Genes with a median q value across all complexes exceeding this threshold are not considered. If NULL, negative correlations will be ignored. By default is 0.8.}
}
\value{
Returns TDAmut object populated with Laplacian scores, p values, and q values for localization of genes passing thresholds of filter_genes
}
\description{
Computes localization of nonsynonymously mutated genes passing thresholds of filter_genes
}
