% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_mut_load.R
\name{compute_mut_load}
\alias{compute_mut_load}
\title{Computes mutational load from input mutation table and assesses localization across grid of nerve complexes}
\usage{
compute_mut_load(
  exp_table,
  mut_table,
  mut_objects,
  hypermut_cutoff = FALSE,
  nonsyn_muts = NA,
  syn_muts = NA,
  graph = FALSE,
  num_cores = 1
)
}
\arguments{
\item{exp_table}{for harmonizing names with mutation data}

\item{mut_table}{mutation data used to compute total mutational load of samples}

\item{mut_objects}{object output by create_mut_object}

\item{hypermut_cutoff}{threshold in log10 scale for downsampling mutations. By default is FALSE.}

\item{nonsyn_muts}{optional input of nonsynonymous mutations in samples. By default is NA.}

\item{syn_muts}{optional input of synonymous mutations in samples. By default is NA.}

\item{graph}{if TRUE, plots histogram of mutational load and heatmap of p values indicating localization across grid of nerve complexes. By default is FALSE.}

\item{num_cores}{number of cores to be used in computation. By default is 1.}
}
\value{
Returns Mutational_Load object containing nonsynonymous mutations, synonymous mutations, and mutational load of samples
}
\description{
Computes mutational load from input mutation table and assesses localization across grid of nerve complexes
}
