% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_mut_object.R
\name{create_mut_object}
\alias{create_mut_object}
\title{Creates network representations of the expression space of a patient cohort across grid of Mapper parameters}
\usage{
create_mut_object(
  exp_table,
  mut_table,
  filter_method = UMAP,
  k = 30,
  min_interval = 10,
  max_interval,
  interval_step = 10,
  min_percent_overlap = 20,
  max_percent_overlap,
  percent_step = 10,
  var_threshold = 4500,
  num_cores = 1
)
}
\arguments{
\item{exp_table}{transcriptomic data from cohort with format samples x genes}

\item{mut_table}{mutation data from cohort in a table with headings Sample, Gene, Mutation, Type}

\item{filter_method}{filter function to create embedding used by TDAmapper. Includes UMAP, KNN, and PCA. By default is set to UMAP.}

\item{k}{if KNN filter function is used, specifies number of nearest neighbors. By default is 30}

\item{min_interval}{minimum number of intervals in codomain of filter function. By default is 10.}

\item{max_interval}{maximum number of intervals in codomain of filter function}

\item{interval_step}{step size between intervals in grid of Mapper parameters. By default is 10.}

\item{min_percent_overlap}{minimum percentage overlap between intervals. By default is 20.}

\item{max_percent_overlap}{maximum percentage overlap between intervals}

\item{percent_step}{step size between percentages in grid of Mapper parameters. By default is 10.}

\item{var_threshold}{number of most variable genes to use from input expression table. By default is 4500.}

\item{num_cores}{number of cores to be used in computation. By default is 1}
}
\value{
Returns a DriverMut_TDA object of nerve complexes computed across range of Mapper parameters.
}
\description{
Creates network representations of the expression space of a patient cohort across grid of Mapper parameters
}
