% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_genes.R
\name{filter_genes}
\alias{filter_genes}
\title{Filter genes considered for compute_localization by mutation frequency, fraction of nonsynonymous mutations, and negative correlations between expression and mutation data}
\usage{
filter_genes(
  TDAmut_object,
  freq_threshold = 0.01,
  top_nonsyn_fraction = 350,
  negative_correlations = TRUE,
  num_permutations = 2000,
  num_cores = 1,
  seed = 121
)
}
\arguments{
\item{TDAmut_object}{object of class TDAmut with expression data, mutation data, nerve complexes}

\item{freq_threshold}{threshold mutation frequency. Genes below this value are not considered in analysis. By default is 0.02}

\item{top_nonsyn_fraction}{number of genes to keep with greatest nonsyn/nonsyn+syn fraction. By default is 350.}

\item{negative_correlations}{if TRUE, assesses negative correlations between mutations rates and expression of a gene. By default is TRUE.}

\item{num_permutations}{if negative_correlations = TRUE, number of permutations used to create null distribution when identifying negative correlations between expression and mutation data. By default is 2000}

\item{num_cores}{number of cores to be used in computation. By default is 1.}

\item{seed}{integer specifying the seed used to initialize the generator of permutations. By default is 121}
}
\value{
Returns a TDAmut object populated with filtered genes to consider in compute_gene_localization. Optionally returns p and q values quantifying negative correlations between expression and mutation profiles of filtered genes.
}
\description{
Filter genes considered for compute_localization by mutation frequency, fraction of nonsynonymous mutations, and negative correlations between expression and mutation data
}
