% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_drivers.R
\name{identify_drivers}
\alias{identify_drivers}
\title{Assesses localization of nonsynonymous mutations across grid of nerve complexes}
\usage{
identify_drivers(
  mut_objects,
  mutload_object,
  freq_threshold = 0.02,
  top_nonsyn_fraction = 350,
  permutations = 5000,
  q_threshold = 0.15,
  num_cores = 1
)
}
\arguments{
\item{mut_objects}{object output by create_mut_object}

\item{mutload_object}{object output by compute_mut_load}

\item{freq_threshold}{threshold mutation frequency. Genes below this value are not considered in analysis. By default is 0.02}

\item{top_nonsyn_fraction}{number of genes to keep with greatest nonsyn/nonsyn+syn fraction. By default is 350.}

\item{permutations}{number of permutations used to create null distribution. By default is 10000}

\item{q_threshold}{threshold q value using Benjamini-Hochberg procedure to control false discovery rate}

\item{num_cores}{number of cores to be used in computation. By default is 1.}
}
\value{
Returns table of significant nonsynonymously mutated genes
}
\description{
Assesses localization of nonsynonymous mutations across grid of nerve complexes
}
