% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_significant_mutations.R
\name{identify_significant_mutations}
\alias{identify_significant_mutations}
\title{Summarizes and identifies significant mutations from filter_genes}
\usage{
identify_significant_mutations(TDAmut_object, q_threshold_localization = 0.15)
}
\arguments{
\item{TDAmut_object}{object of class TDAmut with expression data, mutation data, and nerve complexes}

\item{q_threshold_localization}{threshold q value for gene localization using Benjamini-Hochberg procedure to control false discovery rate}
}
\value{
Returns significant nonsynonymously mutated genes
}
\description{
Summarizes and identifies significant mutations from filter_genes
}
