% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mapper.R
\name{plot_mapper}
\alias{plot_mapper}
\title{Plots features on nerve complexes and embeddings}
\usage{
plot_mapper(
  TDAmut_object,
  type = "none",
  features = NULL,
  which_complexes = NULL,
  colorbar_low = "blue",
  colorbar_high = "red",
  seed = 121
)
}
\arguments{
\item{TDAmut_object}{object of class TDAmut with expression data, mutation data, nerve complexes, mutational load, and mutational load localization data.}

\item{type}{type of data for given features. Can take values 'mutation', 'mutational_load', 'expression', or 'none' to plot complexes with no label. By default is 'none'}

\item{features}{genes of interest in character vector. Only applies if 'mutation' or 'expression' data is selected. By default is NULL}

\item{which_complexes}{vector of interval and percent combinations specificying which nerve complexes to plot, such as c(int1, percent1, int2, percent2 ...). Can plot all complexes by setting to 'All'. By default is NULL which chooses 3 complexes in the middle of the parameter ranges.}

\item{colorbar_low}{color of colorbar for lower-range values. By default is 'blue'}

\item{colorbar_high}{color of colorbar for upper-range values. By default is 'red'}

\item{seed}{seed value when plotting nerve complex. By default is 121.}
}
\value{
Returns plot of nerve complex or embedding with desired features
}
\description{
Plots features on nerve complexes and embeddings
}
