% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mut_load.R
\name{plot_mut_load}
\alias{plot_mut_load}
\title{Plots histogram of mutational load in log10 scale and heatmap of pvalues indicating mutational load localization}
\usage{
plot_mut_load(
  TDAmut_object,
  histogram_breaks = 100,
  pvalue_figs = 3,
  colorbar_low = "red",
  colorbar_mid = "purple",
  colorbar_high = "blue"
)
}
\arguments{
\item{TDAmut_object}{object of class TDAmut with expression data, mutation data, nerve complexes, mutational load, and mutational load localization data.}

\item{histogram_breaks}{number of bins of mutational load histogram. By default is 100.}

\item{pvalue_figs}{number of values to display after decimal in localization heatmap}

\item{colorbar_low}{color of colorbar for low-range values. By default is 'red'}

\item{colorbar_mid}{color of colorbar for mid-range values. By default is 'purple'}

\item{colorbar_high}{color of colorbar for upper-range values. By default is 'blue'}
}
\value{
Returns mutational load histogram and heatmap of mutational load localization across grid of nerve complexes
}
\description{
Plots histogram of mutational load in log10 scale and heatmap of pvalues indicating mutational load localization
}
