% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{class}
\name{TDAmut-class}
\alias{TDAmut-class}
\alias{TDAmut_object}
\title{Class to handle all information in the TDAmut pipeline}
\description{
Class to handle all information in the TDAmut pipeline
}
\section{Slots}{

\describe{
\item{\code{expression_table}}{Expression data data.frame or matrix (create_TDAmut_object)}

\item{\code{mutation_table}}{Mutation data table (create_TDAmut_object)}

\item{\code{nerve_complexes}}{Topological representations as nerve complexes (compute_complexes)}

\item{\code{mapper_intervals}}{Range of intervals used in TDAmapper (compute_complexes)}

\item{\code{mapper_percents}}{Range of percent overlaps used in TDAmapper (compute_complexes)}

\item{\code{filter_embedding}}{Embedding used to create topological representations in TDAmapper (compute_complexes)}

\item{\code{mutation_matrix}}{Reformatted mutation data (create_TDAmut_object)}

\item{\code{nonsyn_mutations}}{Reformatted nonsynonymous mutation data (create_TDAmut_object)}

\item{\code{syn_mutations}}{Reformatted synonymous mutation data (create_TDAmut_object)}

\item{\code{mutational_load}}{Total mutations per sample (create_TDAmut_object)}

\item{\code{mutational_load_localization}}{p and q values indicating significance of mutational load localization across the parameter space (compute_complexes)}

\item{\code{correlations}}{p and q values indicating significance of negative correlation between expression and mutation data of a given gene across the parameter space (filter_genes)}

\item{\code{filtered_genes}}{Genes passing user-defined thresholds on mutational frequency, ratio of nonsynonymous / total mutations, and correlation q values (filter_genes)}

\item{\code{gene_scores}}{p and q values indicating significance of localization of nonsynonymously mutated genes across the parameter space (compute_gene_localization)}

\item{\code{summary_matrix}}{Matrix summarizing the results of the pipelinesin the computation (compute_gene_localization)}
}}

