% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_mut_load.R
\name{compute_mut_load}
\alias{compute_mut_load}
\title{Computes mutational load from input mutation table and assesses localization across grid of nerve complexes}
\usage{
compute_mut_load(
  TDAmut_object,
  hypermut_cutoff = NULL,
  min_mutload = NULL,
  nonsyn_muts = NULL,
  syn_muts = NULL,
  num_permutations = 1000,
  seed = 121
)
}
\arguments{
\item{TDAmut_object}{Object of class TDAmut with expression data, mutation data, and nerve complexes}

\item{hypermut_cutoff}{threshold in log10 scale for downsampling mutations. By default is FALSE.}

\item{min_mutload}{minimum threshold in log10 scale for mutational load of samples. Samples below threshold are saved and to be used to recompute nerve complexes in compute_complexes}

\item{nonsyn_muts}{optional input of nonsynonymous mutations in samples. By default is NA.}

\item{syn_muts}{optional input of synonymous mutations in samples. By default is NA.}

\item{num_permutations}{number of permutations used to create null distribution. By default is 2000}

\item{seed}{integer specifying the seed used to initialize the generator of permutations. By default is 121}
}
\value{
Appends object of class TDAmut with list of pvalues indicating localization of mutational load and 
data frames of nonsynonymous mutations, synonymous mutations, and mutational load of samples
}
\description{
Computes mutational load from input mutation table and assesses localization across grid of nerve complexes
}
