% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mapper.R
\name{plot_mapper}
\alias{plot_mapper}
\title{Plots features on Mapper graph}
\usage{
plot_mapper(
  mapper_obj,
  features = NA,
  feature_table = NA,
  mutational_load = NA
)
}
\arguments{
\item{mapper_obj}{nerve complex on which to plot features}

\item{features}{genes of interest. By default is NA.}

\item{feature_table}{table with feature information such as mutation or expression table. By default is NA.}

\item{mutational_load}{vector of mutational load for each sample. By default is NA.}
}
\value{
Returns labeled plot of nerve complex with given feature
}
\description{
Plots features on Mapper graph
}
