% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{class}
\name{TDAmut-class}
\alias{TDAmut-class}
\alias{TDAmut_object}
\title{Class to handle all information in TDA.TCGA pipeline}
\description{
Class to handle all information in TDA.TCGA pipeline
}
\section{Slots}{

\describe{
\item{\code{expression_table}}{Expression data data.frame or matrix (create_TDAmut_object)}

\item{\code{mutation_table}}{Mutation data table (create_TDAmut_object)}

\item{\code{nerve_complexes}}{Topological representations as nerve complexes (compute_complexes)}

\item{\code{mapper_intervals}}{Range of intervals used in TDAmapper (compute_complexes)}

\item{\code{mapper_percents}}{Range of percent overlaps used in TDAmapper (compute_complexes)}

\item{\code{filter_embedding}}{Embedding used to create topological representations in TDAmapper (compute_complexes)}

\item{\code{mutation_matrix}}{Reformatted mutation data (compute_mut_load)}

\item{\code{nonsyn_mutations}}{Reformatted nonsynonymous mutation data (compute_mut_load)}

\item{\code{syn_mutations}}{Reformatted synonymous mutation data (compute_mut_load)}

\item{\code{mutational_load}}{Total mutations per sample (compute_mut_load)}

\item{\code{min_mutated_samples}}{Samples with mutational load below a user-defined threshold (compute_mut_load)}

\item{\code{mutational_load_localization}}{p and q values indicating significance of mutational load localization across the parameter space (compute_mut_load)}

\item{\code{negative_correlations}}{p and q values indicating significance of correlation between expression and mutation data of a given gene across the parameter space (filter_genes)}

\item{\code{filtered_genes}}{Genes passing user-defined thresholds on mutational frequency and ratio of nonsynonymous / total mutations}

\item{\code{gene_scores}}{p and q values indicating significance of localization of nonsynonymously mutated genes across the parameter space (compute_gene_localization)}

\item{\code{significant_genes}}{Genes passing user-defined median q value threshold across the parameter space (identify_significant_mutations)}
}}

